backend/requirements.txt
fastapi
uvicorn
motor
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
pdfplumber
python-docx
spacy
aiofiles
httpx
python-dotenv
pydantic-settings
fuzzywuzzy
python-Levenshtein
openai
